pyproject.toml
[project]
name = "eopf-mcp"
version = "0.1.0"
description = "MCP server for EOPF STAC catalog - enables LLMs to query Earth Observation data"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.84.0",
    "bokeh>=3.9.0",
    "bottleneck==1.4.2",
    "cartopy>=0.25.0",
    "dask>=2026.1.2",
    "fastapi>=0.115.0",
    "folium>=0.20.0",
    "fsspec>=2026.2.0",
    "gdal==3.8.4",
    "geopandas>=1.1.3",
    "gradio>=6.9.0",
    "graphviz>=0.21",
    "holoviews>=1.14.9",
    "hvplot>=0.8.1",
    "ipykernel>=7.2.0",
    "mapclassify>=2.10.0",
    "matplotlib==3.10.0",
    "matplotlib-inline>=0.2.1",
    "mcp>=1.26.0",
    "myst-nb-bokeh>=2024.1.0.post1",
    "mystmd>=1.7.0",
    "nbformat>=5.10.4",
    "netcdf4>=1.7.4",
    "numpy>=2.4.3",
    "onnxruntime>=1.24.3",
    "openeo>=0.44.0",
    "pystac-client>=0.9.0",
    "pyyaml>=6.0.3",
    "rioxarray>=0.21.0",
    "s3fs>=2026.2.0",
    "scikit-image>=0.26.0",
    "tqdm>=4.67.3",
    "uvicorn>=0.42.0",
    "xarray>=2025.1.1",
    "xarray-eopf>=0.2.4",
    "xcube-core<1.13",
    "xcube-eopf>=0.3.0",
    "zarr<3",
]
